requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
Pillow==9.5.0
Pillow==10.0.1
roboflow==1.1.6
